[MIG] sale_stock_operating_unit to v10.
