+++ sale_stock_operating_unit/__manifest__.py
@@ -1,18 +1,18 @@
 # -*- coding: utf-8 -*-
-# © 2015 Eficent Business and IT Consulting Services S.L. -
+# © 2015-17 Eficent Business and IT Consulting Services S.L. -
 # Jordi Ballester Alomar
-# © 2015 Serpent Consulting Services Pvt. Ltd. - Sudhir Arya
+# © 2015-17 Serpent Consulting Services Pvt. Ltd. - Sudhir Arya
 # License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl.html).
 
 {
     "name": "Operating Unit in Sales Stock",
     "summary": "An operating unit (OU) is an organizational entity part of a "
                "company",
-    "version": "9.0.1.0.0",
+    "version": "10.0.1.0.0",
     "author": "Eficent, Serpent Consulting Services Pvt. Ltd., "
               "Odoo Community Association (OCA)",
     "license": "LGPL-3",
-    "website": "http://www.eficent.com",
+    "website": "https://github.com/OCA/operating-unit",
     "category": "Sales Management",
     "depends": [
         "sale_stock",
