+++ sale_stock_operating_unit/tests/test_sale_stock_operating_unit.py
@@ -1,10 +1,10 @@
 # -*- coding: utf-8 -*-
-# © 2015 Eficent Business and IT Consulting Services S.L. -
+# © 2015-17 Eficent Business and IT Consulting Services S.L. -
 # Jordi Ballester Alomar
-# © 2015 Serpent Consulting Services Pvt. Ltd. - Sudhir Arya
+# © 2015-17 Serpent Consulting Services Pvt. Ltd. - Sudhir Arya
 # License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl.html).
 
-from openerp.addons.sale_operating_unit.tests import test_sale_operating_unit
+from odoo.addons.sale_operating_unit.tests import test_sale_operating_unit
 
 
 class TestSaleStockOperatingUnit(test_sale_operating_unit.
